@@ -1,14 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
     <ItemGroup>
       <PackageReference Include="FluentAssertions" Version="5.10.3" />
-      <PackageReference Include="Microsoft.AspNetCore.App" />
-      <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="2.2.0" />
+      <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.0" />
       <PackageReference Include="Nunit" Version="3.11.0" />
       <PackageReference Include="NUnit3TestAdapter" Version="3.11.0" />
       <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
