@@ -1,21 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
 
         <IsPackable>false</IsPackable>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="AutoFixture" Version="4.11.0" />
-        <PackageReference Include="AutoFixture.NUnit3" Version="4.11.0" />
+        <PackageReference Include="AutoFixture" Version="4.12.0" />
+        <PackageReference Include="AutoFixture.NUnit3" Version="4.12.0" />
         <PackageReference Include="FluentAssertions" Version="5.10.3" />
         <PackageReference Include="Lucene.Net" Version="4.8.0-beta00008" />
-        <PackageReference Include="Moq" Version="4.14.1" />
+        <PackageReference Include="Moq" Version="4.14.5" />
         <PackageReference Include="nunit" Version="3.12.0" />
-        <PackageReference Include="NUnit3TestAdapter" Version="3.13.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
-        <PackageReference Include="SFA.DAS.Testing.AutoFixture" Version="3.0.106" />
+        <PackageReference Include="NUnit3TestAdapter" Version="3.16.1">
+          <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
+        <PackageReference Include="SFA.DAS.Testing.AutoFixture" Version="3.0.169" />
     </ItemGroup>
 
     <ItemGroup>
