@@ -1,13 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
     <ItemGroup>
         <PackageReference Include="Lucene.Net" Version="4.8.0-beta00008" />
         <PackageReference Include="Lucene.Net.Analysis.Phonetic" Version="4.8.0-beta00008" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="2.2.0" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.0" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="3.1.5" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.5" />
     </ItemGroup>
     <ItemGroup>
       <ProjectReference Include="..\SFA.DAS.Courses.Domain\SFA.DAS.Courses.Domain.csproj" />
