@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using Microsoft.OpenApi.Models;
 using SFA.DAS.Configuration.AzureTableStorage;
@@ -82,7 +83,7 @@ namespace SFA.DAS.Courses.Api
                     {
                         o.Conventions.Add(new AuthorizeControllerModelConvention());
                     }
-                }).SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                }).SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
             
             services.AddApplicationInsightsTelemetry(_configuration["APPINSIGHTS_INSTRUMENTATIONKEY"]);
 
@@ -96,7 +97,7 @@ namespace SFA.DAS.Courses.Api
             indexBuilder.Build();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -106,21 +107,21 @@ namespace SFA.DAS.Courses.Api
             app.UseAuthentication();    
         
             app.UseHealthChecks();
-            
-            app.UseMvc(routes =>
+
+            app.UseRouting();
+            app.UseEndpoints(builder =>
             {
-                routes.MapRoute(
+                builder.MapControllerRoute(
                     name: "default",
-                    template: "api/{controller=Standards}/{action=Index}/{id?}");
+                    pattern: "api/{controller=Standards}/{action=Index}/{id?}");
             });
-            
+
             app.UseSwagger();
             app.UseSwaggerUI(c =>
             {
                 c.SwaggerEndpoint("/swagger/v1/swagger.json", "CoursesAPI");
                 c.RoutePrefix = string.Empty;
             });
-            
         }
         
         private bool ConfigurationIsLocalOrDev()
